Social_distancing_with_face_mask/requirements.txt
absl-py==0.13.0
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bcrypt==3.2.0
bidict==0.21.2
bleach==4.1.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cmake==3.21.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dnspython==2.1.0
email-validator==1.1.2
entrypoints==0.3
face-detection @ git+https://github.com/hukkelas/DSFD-Pytorch-Inference.git@2bdd997d785e20ea39a911e9b3c451b7cdd3b152
face-recognition==1.4.0
face-recognition-models==0.3.0
flasgger==0.9.4
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.39.0
gunicorn==19.9.0
h5py==2.10.0
idna==3.2
imutils==0.5.4
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.10.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.18.5
oauthlib==3.1.1
opencv-python==4.3.0.36
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.7.0
pygal==2.4.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.2.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.23
streamlit==0.88.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-gpu==2.3.0
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.15.5
wget==3.2
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
